+++ CHANGELOG.md
@@ -90,7 +90,8 @@ CORE
 	/ gstreamer: faster reload by default and optional asynchronous load
 	/ ofVideoPlayer::loadMovie() -> load()
 	/ ofVideoGrabber::initGrabber() -> setup()
-
+### types
+    + ofRectangle: added `setSize` method.
 
 PLATFORM SPECIFIC
 -----------------